Added Signin and login feature

+++ dividend_calculator/app.py
@@ -1,15 +1,18 @@
 from flask import Blueprint, render_template, request, redirect, url_for
+from flask_login import login_required, current_user
 from .extensions import db
 from .dividend_calculator import DividendCalculator
 from .models import Stock
 import yfinance as yf
 
 routes = Blueprint('routes', __name__)
-    
+
 @routes.route('/')
 def index():
-    stocks = Stock.query.all()
-
+    if  current_user.is_authenticated:
+        stocks = Stock.query.filter_by(user_id=current_user.id).all()
+    else:
+        stocks = []
     total_quarterly_payout = 0
     total_yearly_payout = 0
     stock_details = []
@@ -31,35 +34,37 @@ def index():
 
 
 @routes.route('/add_stock', methods=['POST'])
+@login_required
 def add_stock():
     ticker = request.form['ticker']
     shares = float(request.form['shares'])
 
-    existing_stock = Stock.query.filter_by(ticker=ticker).first()
+    existing_stock = Stock.query.filter_by(ticker=ticker, user_id=current_user.id).first()
 
     if existing_stock:
         existing_stock.shares += shares
     else:
-        new_stock = Stock(ticker=ticker, shares=shares)
+        new_stock = Stock(ticker=ticker, shares=shares, user_id=current_user.id)
         db.session.add(new_stock)
 
     db.session.commit()
     return redirect(url_for('routes.index'))
 
 @routes.route('/delete_stock/<string:ticker>', methods=['GET'])
+@login_required
 def delete_stock(ticker):
-    Stock.query.filter_by(ticker=ticker).delete()
+    Stock.query.filter_by(ticker=ticker, user_id=current_user.id).delete()
     db.session.commit()
     return redirect(url_for('routes.index'))
 
-
 @routes.route('/edit_stock', methods=['POST'])
+@login_required
 def edit_stock():
     old_ticker = request.form['old_ticker']
     new_ticker = request.form['new_ticker']
     new_shares = request.form['new_shares']
 
-    stock = Stock.query.filter_by(ticker=old_ticker).first()
+    stock = Stock.query.filter_by(ticker=old_ticker, user_id=current_user.id).first()
 
     if stock:
         if new_ticker:
@@ -72,9 +77,11 @@ def edit_stock():
 
 
 @routes.route('/get_tickers')
+@login_required
 def get_tickers():
-    tickers = [stock.ticker for stock in Stock.query.with_entities(Stock.ticker).distinct()]
+    tickers = [stock.ticker for stock in Stock.query.filter_by(user_id=current_user.id).with_entities(Stock.ticker).distinct()]
     return {'tickers': tickers}
+
